487: Applying refactoring to account scheduled payments and standing order packages
Applying the test refactoring to allow logic to be reused to call different API versions

Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/487
