+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/scheduled/payments/api/v3_1_8/GetScheduledPayments.kt
@@ -1,0 +1,31 @@
+package com.forgerock.uk.openbanking.tests.functional.account.scheduled.payments.api.v3_1_8
+
+import assertk.assertThat
+import assertk.assertions.isNotEmpty
+import assertk.assertions.isNotNull
+import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
+import com.forgerock.uk.openbanking.support.account.AccountRS
+import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
+import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code
+import uk.org.openbanking.datamodel.account.OBReadScheduledPayment3
+
+class GetScheduledPayments(version: OBVersion, tppResource: CreateTppCallback.TppResource) :
+    BaseAccountApi3_1_8(version, tppResource) {
+
+    fun shouldGetScheduledPaymentsTest() {
+        // Given
+        val permissions = listOf(OBExternalPermissions1Code.READSCHEDULEDPAYMENTSDETAIL)
+        val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
+
+        // When
+        val result = AccountRS().getAccountsData<OBReadScheduledPayment3>(
+            accountsApiLinks.GetScheduledPayments,
+            accessToken
+        )
+
+        // Then
+        assertThat(result).isNotNull()
+        assertThat(result.data.scheduledPayment).isNotEmpty()
+    }
+}
