+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/scheduled/payments/junit/v3_1_8/GetScheduledPaymentsTest.kt
@@ -1,0 +1,21 @@
+package com.forgerock.uk.openbanking.tests.functional.account.scheduled.payments.junit.v3_1_8
+
+import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
+import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
+import com.forgerock.uk.openbanking.tests.functional.account.scheduled.payments.api.v3_1_8.GetScheduledPayments
+import org.junit.jupiter.api.Test
+
+class GetScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
+    @EnabledIfVersion(
+        type = "accounts",
+        apiVersion = "v3.1.8",
+        operations = ["CreateAccountAccessConsent", "GetScheduledPayments"],
+        apis = ["scheduled-payments"],
+        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
+    )
+    @Test
+    fun shouldGetScheduledPayments_v3_1_8() {
+        GetScheduledPayments(OBVersion.v3_1_8, tppResource).shouldGetScheduledPaymentsTest()
+    }
+}
