+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/scheduled/payments/legacy/LegacyGetAccountScheduledPaymentsTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.scheduled.payments
+package com.forgerock.uk.openbanking.tests.functional.account.scheduled.payments.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -7,31 +7,33 @@ import com.forgerock.securebanking.framework.configuration.psu
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.uk.openbanking.support.account.AccountAS
-import com.forgerock.uk.openbanking.support.account.AccountFactory.Companion.obReadConsent1
+import com.forgerock.uk.openbanking.support.account.AccountFactory
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_4
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import org.junit.jupiter.api.Test
+import uk.org.openbanking.datamodel.account.*
+import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READACCOUNTSDETAIL
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READSCHEDULEDPAYMENTSDETAIL
-import uk.org.openbanking.datamodel.account.OBReadConsentResponse1
-import uk.org.openbanking.datamodel.account.OBReadScheduledPayment2
-import uk.org.openbanking.datamodel.account.OBReadScheduledPayment3
 
-class GetScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetAccountScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1",
-        operations = ["CreateAccountAccessConsent", "GetScheduledPayments"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountScheduledPayments"],
         apis = ["scheduled-payments"],
         compatibleVersions = ["v.3.0"]
     )
     @Test
-    fun shouldGetScheduledPayments_v3_1() {
+    fun shouldGetAccountScheduledPayments_v3_1() {
         // Given
-        val consentRequest = obReadConsent1(listOf(READSCHEDULEDPAYMENTSDETAIL))
+        val consentRequest = OBReadConsent1().data(
+            OBReadData1()
+                .permissions(listOf(READACCOUNTSDETAIL, READSCHEDULEDPAYMENTSDETAIL))
+        )
+            .risk(OBRisk2())
         val consent = AccountRS().consent<OBReadConsentResponse1>(
             accountAndTransaction3_1.Links.links.CreateAccountAccessConsent,
             consentRequest,
@@ -43,11 +45,14 @@ class GetScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
             psu,
             tppResource.tpp
         )
+        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1.Links.links.GetAccounts, accessToken)
 
         // When
-        val result = AccountRS().getAccountsData<OBReadScheduledPayment3>(
-            accountAndTransaction3_1.Links.links.GetScheduledPayments,
-            accessToken
+        val result = AccountRS().getAccountsData<OBReadScheduledPayment2>(
+            AccountFactory.urlWithAccountId(
+                accountAndTransaction3_1.Links.links.GetAccountScheduledPayments,
+                accountId
+            ), accessToken
         )
 
         // Then
@@ -58,14 +63,18 @@ class GetScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1.2",
-        operations = ["CreateAccountAccessConsent", "GetScheduledPayments"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountScheduledPayments"],
         apis = ["scheduled-payments"],
         compatibleVersions = ["v.3.1.1"]
     )
     @Test
-    fun shouldGetScheduledPayments_v3_1_2() {
+    fun shouldGetAccountScheduledPayments_v3_1_2() {
         // Given
-        val consentRequest = obReadConsent1(listOf(READSCHEDULEDPAYMENTSDETAIL))
+        val consentRequest = OBReadConsent1().data(
+            OBReadData1()
+                .permissions(listOf(READACCOUNTSDETAIL, READSCHEDULEDPAYMENTSDETAIL))
+        )
+            .risk(OBRisk2())
         val consent = AccountRS().consent<OBReadConsentResponse1>(
             accountAndTransaction3_1_2.Links.links.CreateAccountAccessConsent,
             consentRequest,
@@ -77,11 +86,14 @@ class GetScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
             psu,
             tppResource.tpp
         )
+        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_2.Links.links.GetAccounts, accessToken)
 
         // When
         val result = AccountRS().getAccountsData<OBReadScheduledPayment2>(
-            accountAndTransaction3_1_2.Links.links.GetScheduledPayments,
-            accessToken
+            AccountFactory.urlWithAccountId(
+                accountAndTransaction3_1_2.Links.links.GetAccountScheduledPayments,
+                accountId
+            ), accessToken
         )
 
         // Then
@@ -92,14 +104,18 @@ class GetScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1.4",
-        operations = ["CreateAccountAccessConsent", "GetScheduledPayments"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountScheduledPayments"],
         apis = ["scheduled-payments"],
         compatibleVersions = ["v.3.1.3"]
     )
     @Test
-    fun shouldGetScheduledPayments_v3_1_4() {
+    fun shouldGetAccountScheduledPayments_v3_1_4() {
         // Given
-        val consentRequest = obReadConsent1(listOf(READSCHEDULEDPAYMENTSDETAIL))
+        val consentRequest = OBReadConsent1().data(
+            OBReadData1()
+                .permissions(listOf(READACCOUNTSDETAIL, READSCHEDULEDPAYMENTSDETAIL))
+        )
+            .risk(OBRisk2())
         val consent = AccountRS().consent<OBReadConsentResponse1>(
             accountAndTransaction3_1_4.Links.links.CreateAccountAccessConsent,
             consentRequest,
@@ -111,45 +127,14 @@ class GetScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
             psu,
             tppResource.tpp
         )
+        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_4.Links.links.GetAccounts, accessToken)
 
         // When
         val result = AccountRS().getAccountsData<OBReadScheduledPayment3>(
-            accountAndTransaction3_1_4.Links.links.GetScheduledPayments,
-            accessToken
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data.scheduledPayment).isNotEmpty()
-    }
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetScheduledPayments"],
-        apis = ["scheduled-payments"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldGetScheduledPayments_v3_1_8() {
-        // Given
-        val consentRequest = obReadConsent1(listOf(READSCHEDULEDPAYMENTSDETAIL))
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-
-        // When
-        val result = AccountRS().getAccountsData<OBReadScheduledPayment3>(
-            accountAndTransaction3_1_8.Links.links.GetScheduledPayments,
-            accessToken
+            AccountFactory.urlWithAccountId(
+                accountAndTransaction3_1_4.Links.links.GetAccountScheduledPayments,
+                accountId
+            ), accessToken
         )
 
         // Then
