+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/scheduled/payments/legacy/LegacyGetScheduledPaymentsTest.kt
@@ -1,0 +1,124 @@
+package com.forgerock.uk.openbanking.tests.functional.account.scheduled.payments.legacy
+
+import assertk.assertThat
+import assertk.assertions.isNotEmpty
+import assertk.assertions.isNotNull
+import com.forgerock.securebanking.framework.configuration.psu
+import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
+import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
+import com.forgerock.uk.openbanking.support.account.AccountAS
+import com.forgerock.uk.openbanking.support.account.AccountFactory.Companion.obReadConsent1
+import com.forgerock.uk.openbanking.support.account.AccountRS
+import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1
+import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
+import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_4
+import org.junit.jupiter.api.Test
+import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READSCHEDULEDPAYMENTSDETAIL
+import uk.org.openbanking.datamodel.account.OBReadConsentResponse1
+import uk.org.openbanking.datamodel.account.OBReadScheduledPayment2
+import uk.org.openbanking.datamodel.account.OBReadScheduledPayment3
+
+class LegacyGetScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
+
+    @EnabledIfVersion(
+        type = "accounts",
+        apiVersion = "v3.1",
+        operations = ["CreateAccountAccessConsent", "GetScheduledPayments"],
+        apis = ["scheduled-payments"],
+        compatibleVersions = ["v.3.0"]
+    )
+    @Test
+    fun shouldGetScheduledPayments_v3_1() {
+        // Given
+        val consentRequest = obReadConsent1(listOf(READSCHEDULEDPAYMENTSDETAIL))
+        val consent = AccountRS().consent<OBReadConsentResponse1>(
+            accountAndTransaction3_1.Links.links.CreateAccountAccessConsent,
+            consentRequest,
+            tppResource.tpp
+        )
+        val accessToken = AccountAS().getAccessToken(
+            consent.data.consentId,
+            tppResource.tpp.registrationResponse,
+            psu,
+            tppResource.tpp
+        )
+
+        // When
+        val result = AccountRS().getAccountsData<OBReadScheduledPayment3>(
+            accountAndTransaction3_1.Links.links.GetScheduledPayments,
+            accessToken
+        )
+
+        // Then
+        assertThat(result).isNotNull()
+        assertThat(result.data.scheduledPayment).isNotEmpty()
+    }
+
+    @EnabledIfVersion(
+        type = "accounts",
+        apiVersion = "v3.1.2",
+        operations = ["CreateAccountAccessConsent", "GetScheduledPayments"],
+        apis = ["scheduled-payments"],
+        compatibleVersions = ["v.3.1.1"]
+    )
+    @Test
+    fun shouldGetScheduledPayments_v3_1_2() {
+        // Given
+        val consentRequest = obReadConsent1(listOf(READSCHEDULEDPAYMENTSDETAIL))
+        val consent = AccountRS().consent<OBReadConsentResponse1>(
+            accountAndTransaction3_1_2.Links.links.CreateAccountAccessConsent,
+            consentRequest,
+            tppResource.tpp
+        )
+        val accessToken = AccountAS().getAccessToken(
+            consent.data.consentId,
+            tppResource.tpp.registrationResponse,
+            psu,
+            tppResource.tpp
+        )
+
+        // When
+        val result = AccountRS().getAccountsData<OBReadScheduledPayment2>(
+            accountAndTransaction3_1_2.Links.links.GetScheduledPayments,
+            accessToken
+        )
+
+        // Then
+        assertThat(result).isNotNull()
+        assertThat(result.data.scheduledPayment).isNotEmpty()
+    }
+
+    @EnabledIfVersion(
+        type = "accounts",
+        apiVersion = "v3.1.4",
+        operations = ["CreateAccountAccessConsent", "GetScheduledPayments"],
+        apis = ["scheduled-payments"],
+        compatibleVersions = ["v.3.1.3"]
+    )
+    @Test
+    fun shouldGetScheduledPayments_v3_1_4() {
+        // Given
+        val consentRequest = obReadConsent1(listOf(READSCHEDULEDPAYMENTSDETAIL))
+        val consent = AccountRS().consent<OBReadConsentResponse1>(
+            accountAndTransaction3_1_4.Links.links.CreateAccountAccessConsent,
+            consentRequest,
+            tppResource.tpp
+        )
+        val accessToken = AccountAS().getAccessToken(
+            consent.data.consentId,
+            tppResource.tpp.registrationResponse,
+            psu,
+            tppResource.tpp
+        )
+
+        // When
+        val result = AccountRS().getAccountsData<OBReadScheduledPayment3>(
+            accountAndTransaction3_1_4.Links.links.GetScheduledPayments,
+            accessToken
+        )
+
+        // Then
+        assertThat(result).isNotNull()
+        assertThat(result.data.scheduledPayment).isNotEmpty()
+    }
+}
