+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/standing/orders/api/v3_1_8/GetStandingOrders.kt
@@ -1,0 +1,32 @@
+package com.forgerock.uk.openbanking.tests.functional.account.standing.orders.api.v3_1_8
+
+import assertk.assertThat
+import assertk.assertions.isNotEmpty
+import assertk.assertions.isNotNull
+import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
+import com.forgerock.uk.openbanking.support.account.AccountRS
+import com.forgerock.uk.openbanking.tests.functional.account.access.BaseAccountApi3_1_8
+import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code
+import uk.org.openbanking.datamodel.account.OBReadStandingOrder6
+
+class GetStandingOrders(version: OBVersion, tppResource: CreateTppCallback.TppResource) :
+    BaseAccountApi3_1_8(version, tppResource) {
+
+    fun shouldGetStandingOrdersTest() {
+        // Given
+        val permissions =
+            listOf(OBExternalPermissions1Code.READACCOUNTSDETAIL, OBExternalPermissions1Code.READSTANDINGORDERSDETAIL)
+        val (_, accessToken) = accountAccessConsentApi.createConsentAndGetAccessToken(permissions)
+
+        // When
+        val result = AccountRS().getAccountsData<OBReadStandingOrder6>(
+            accountsApiLinks.GetStandingOrders,
+            accessToken
+        )
+
+        // Then
+        assertThat(result).isNotNull()
+        assertThat(result.data.standingOrder).isNotEmpty()
+    }
+}
