+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/standing/orders/legacy/LegacyGetAccountStandingOrdersTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.scheduled.payments
+package com.forgerock.uk.openbanking.tests.functional.account.standing.orders.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -7,32 +7,30 @@ import com.forgerock.securebanking.framework.configuration.psu
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.uk.openbanking.support.account.AccountAS
-import com.forgerock.uk.openbanking.support.account.AccountFactory
+import com.forgerock.uk.openbanking.support.account.AccountFactory.Companion.urlWithAccountId
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_4
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.account.*
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READACCOUNTSDETAIL
-import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READSCHEDULEDPAYMENTSDETAIL
 
-class GetAccountScheduledPaymentsTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetAccountStandingOrdersTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountScheduledPayments"],
-        apis = ["scheduled-payments"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountStandingOrders"],
+        apis = ["standing-orders"],
         compatibleVersions = ["v.3.0"]
     )
     @Test
-    fun shouldGetAccountScheduledPayments_v3_1() {
+    fun shouldGetAccountStandingOrders_v3_1() {
         // Given
         val consentRequest = OBReadConsent1().data(
             OBReadData1()
-                .permissions(listOf(READACCOUNTSDETAIL, READSCHEDULEDPAYMENTSDETAIL))
+                .permissions(listOf(READACCOUNTSDETAIL, OBExternalPermissions1Code.READSTANDINGORDERSDETAIL))
         )
             .risk(OBRisk2())
         val consent = AccountRS().consent<OBReadConsentResponse1>(
@@ -49,31 +47,31 @@ class GetAccountScheduledPaymentsTest(val tppResource: CreateTppCallback.TppReso
         val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1.Links.links.GetAccounts, accessToken)
 
         // When
-        val result = AccountRS().getAccountsData<OBReadScheduledPayment2>(
-            AccountFactory.urlWithAccountId(
-                accountAndTransaction3_1.Links.links.GetAccountScheduledPayments,
+        val result = AccountRS().getAccountsData<OBReadStandingOrder4>(
+            urlWithAccountId(
+                accountAndTransaction3_1.Links.links.GetAccountStandingOrders,
                 accountId
             ), accessToken
         )
 
         // Then
         assertThat(result).isNotNull()
-        assertThat(result.data.scheduledPayment).isNotEmpty()
+        assertThat(result.data.standingOrder).isNotEmpty()
     }
 
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1.2",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountScheduledPayments"],
-        apis = ["scheduled-payments"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountStandingOrders"],
+        apis = ["standing-orders"],
         compatibleVersions = ["v.3.1.1"]
     )
     @Test
-    fun shouldGetAccountScheduledPayments_v3_1_2() {
+    fun shouldGetAccountStandingOrders_v3_1_2() {
         // Given
         val consentRequest = OBReadConsent1().data(
             OBReadData1()
-                .permissions(listOf(READACCOUNTSDETAIL, READSCHEDULEDPAYMENTSDETAIL))
+                .permissions(listOf(READACCOUNTSDETAIL, OBExternalPermissions1Code.READSTANDINGORDERSDETAIL))
         )
             .risk(OBRisk2())
         val consent = AccountRS().consent<OBReadConsentResponse1>(
@@ -90,31 +88,31 @@ class GetAccountScheduledPaymentsTest(val tppResource: CreateTppCallback.TppReso
         val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_2.Links.links.GetAccounts, accessToken)
 
         // When
-        val result = AccountRS().getAccountsData<OBReadScheduledPayment2>(
-            AccountFactory.urlWithAccountId(
-                accountAndTransaction3_1_2.Links.links.GetAccountScheduledPayments,
+        val result = AccountRS().getAccountsData<OBReadStandingOrder5>(
+            urlWithAccountId(
+                accountAndTransaction3_1_2.Links.links.GetAccountStandingOrders,
                 accountId
             ), accessToken
         )
 
         // Then
         assertThat(result).isNotNull()
-        assertThat(result.data.scheduledPayment).isNotEmpty()
+        assertThat(result.data.standingOrder).isNotEmpty()
     }
 
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1.4",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountScheduledPayments"],
-        apis = ["scheduled-payments"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountStandingOrders"],
+        apis = ["standing-orders"],
         compatibleVersions = ["v.3.1.3"]
     )
     @Test
-    fun shouldGetAccountScheduledPayments_v3_1_4() {
+    fun shouldGetAccountStandingOrders_v3_1_4() {
         // Given
         val consentRequest = OBReadConsent1().data(
             OBReadData1()
-                .permissions(listOf(READACCOUNTSDETAIL, READSCHEDULEDPAYMENTSDETAIL))
+                .permissions(listOf(READACCOUNTSDETAIL, OBExternalPermissions1Code.READSTANDINGORDERSDETAIL))
         )
             .risk(OBRisk2())
         val consent = AccountRS().consent<OBReadConsentResponse1>(
@@ -131,56 +129,15 @@ class GetAccountScheduledPaymentsTest(val tppResource: CreateTppCallback.TppReso
         val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_4.Links.links.GetAccounts, accessToken)
 
         // When
-        val result = AccountRS().getAccountsData<OBReadScheduledPayment3>(
-            AccountFactory.urlWithAccountId(
-                accountAndTransaction3_1_4.Links.links.GetAccountScheduledPayments,
+        val result = AccountRS().getAccountsData<OBReadStandingOrder6>(
+            urlWithAccountId(
+                accountAndTransaction3_1_4.Links.links.GetAccountStandingOrders,
                 accountId
             ), accessToken
         )
 
         // Then
         assertThat(result).isNotNull()
-        assertThat(result.data.scheduledPayment).isNotEmpty()
-    }
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountScheduledPayments"],
-        apis = ["scheduled-payments"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldGetAccountScheduledPayments_v3_1_8() {
-        // Given
-        val consentRequest = OBReadConsent1().data(
-            OBReadData1()
-                .permissions(listOf(READACCOUNTSDETAIL, READSCHEDULEDPAYMENTSDETAIL))
-        )
-            .risk(OBRisk2())
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_8.Links.links.GetAccounts, accessToken)
-
-        // When
-        val result = AccountRS().getAccountsData<OBReadScheduledPayment3>(
-            AccountFactory.urlWithAccountId(
-                accountAndTransaction3_1_8.Links.links.GetAccountScheduledPayments,
-                accountId
-            ), accessToken
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data.scheduledPayment).isNotEmpty()
+        assertThat(result.data.standingOrder).isNotEmpty()
     }
 }
