+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/account/standing/orders/legacy/LegacyGetStandingOrdersTest.kt
@@ -1,4 +1,4 @@
-package com.forgerock.uk.openbanking.tests.functional.account.standing.orders
+package com.forgerock.uk.openbanking.tests.functional.account.standing.orders.legacy
 
 import assertk.assertThat
 import assertk.assertions.isNotEmpty
@@ -7,27 +7,25 @@ import com.forgerock.securebanking.framework.configuration.psu
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.uk.openbanking.support.account.AccountAS
-import com.forgerock.uk.openbanking.support.account.AccountFactory.Companion.urlWithAccountId
 import com.forgerock.uk.openbanking.support.account.AccountRS
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1
 import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_2
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_4
-import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_8
+import com.forgerock.uk.openbanking.support.discovery.accountAndTransaction3_1_6
 import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.account.*
 import uk.org.openbanking.datamodel.account.OBExternalPermissions1Code.READACCOUNTSDETAIL
 
-class GetAccountStandingOrdersTest(val tppResource: CreateTppCallback.TppResource) {
+class LegacyGetStandingOrdersTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountStandingOrders"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetStandingOrders"],
         apis = ["standing-orders"],
         compatibleVersions = ["v.3.0"]
     )
     @Test
-    fun shouldGetAccountStandingOrders_v3_1() {
+    fun shouldGetStandingOrders_v3_1() {
         // Given
         val consentRequest = OBReadConsent1().data(
             OBReadData1()
@@ -45,14 +43,11 @@ class GetAccountStandingOrdersTest(val tppResource: CreateTppCallback.TppResourc
             psu,
             tppResource.tpp
         )
-        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1.Links.links.GetAccounts, accessToken)
 
         // When
         val result = AccountRS().getAccountsData<OBReadStandingOrder4>(
-            urlWithAccountId(
-                accountAndTransaction3_1.Links.links.GetAccountStandingOrders,
-                accountId
-            ), accessToken
+            accountAndTransaction3_1.Links.links.GetStandingOrders,
+            accessToken
         )
 
         // Then
@@ -63,12 +58,12 @@ class GetAccountStandingOrdersTest(val tppResource: CreateTppCallback.TppResourc
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1.2",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountStandingOrders"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetStandingOrders"],
         apis = ["standing-orders"],
         compatibleVersions = ["v.3.1.1"]
     )
     @Test
-    fun shouldGetAccountStandingOrders_v3_1_2() {
+    fun shouldGetStandingOrders_v3_1_2() {
         // Given
         val consentRequest = OBReadConsent1().data(
             OBReadData1()
@@ -86,14 +81,11 @@ class GetAccountStandingOrdersTest(val tppResource: CreateTppCallback.TppResourc
             psu,
             tppResource.tpp
         )
-        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_2.Links.links.GetAccounts, accessToken)
 
         // When
         val result = AccountRS().getAccountsData<OBReadStandingOrder5>(
-            urlWithAccountId(
-                accountAndTransaction3_1_2.Links.links.GetAccountStandingOrders,
-                accountId
-            ), accessToken
+            accountAndTransaction3_1.Links.links.GetStandingOrders,
+            accessToken
         )
 
         // Then
@@ -104,12 +96,12 @@ class GetAccountStandingOrdersTest(val tppResource: CreateTppCallback.TppResourc
     @EnabledIfVersion(
         type = "accounts",
         apiVersion = "v3.1.4",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountStandingOrders"],
+        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetStandingOrders"],
         apis = ["standing-orders"],
         compatibleVersions = ["v.3.1.3"]
     )
     @Test
-    fun shouldGetAccountStandingOrders_v3_1_4() {
+    fun shouldGetStandingOrders_v3_1_4() {
         // Given
         val consentRequest = OBReadConsent1().data(
             OBReadData1()
@@ -117,7 +109,7 @@ class GetAccountStandingOrdersTest(val tppResource: CreateTppCallback.TppResourc
         )
             .risk(OBRisk2())
         val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_4.Links.links.CreateAccountAccessConsent,
+            accountAndTransaction3_1_6.Links.links.CreateAccountAccessConsent,
             consentRequest,
             tppResource.tpp
         )
@@ -127,56 +119,11 @@ class GetAccountStandingOrdersTest(val tppResource: CreateTppCallback.TppResourc
             psu,
             tppResource.tpp
         )
-        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_4.Links.links.GetAccounts, accessToken)
 
         // When
-        val result = AccountRS().getAccountsData<OBReadStandingOrder6>(
-            urlWithAccountId(
-                accountAndTransaction3_1_4.Links.links.GetAccountStandingOrders,
-                accountId
-            ), accessToken
-        )
-
-        // Then
-        assertThat(result).isNotNull()
-        assertThat(result.data.standingOrder).isNotEmpty()
-    }
-
-
-    @EnabledIfVersion(
-        type = "accounts",
-        apiVersion = "v3.1.8",
-        operations = ["CreateAccountAccessConsent", "GetAccounts", "GetAccountStandingOrders"],
-        apis = ["standing-orders"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldGetAccountStandingOrders_v3_1_8() {
-        // Given
-        val consentRequest = OBReadConsent1().data(
-            OBReadData1()
-                .permissions(listOf(READACCOUNTSDETAIL, OBExternalPermissions1Code.READSTANDINGORDERSDETAIL))
-        )
-            .risk(OBRisk2())
-        val consent = AccountRS().consent<OBReadConsentResponse1>(
-            accountAndTransaction3_1_8.Links.links.CreateAccountAccessConsent,
-            consentRequest,
-            tppResource.tpp
-        )
-        val accessToken = AccountAS().getAccessToken(
-            consent.data.consentId,
-            tppResource.tpp.registrationResponse,
-            psu,
-            tppResource.tpp
-        )
-        val accountId = AccountRS().getFirstAccountId(accountAndTransaction3_1_8.Links.links.GetAccounts, accessToken)
-
-        // When
-        val result = AccountRS().getAccountsData<OBReadStandingOrder6>(
-            urlWithAccountId(
-                accountAndTransaction3_1_8.Links.links.GetAccountStandingOrders,
-                accountId
-            ), accessToken
+        val result = AccountRS().getAccountsData<OBReadStandingOrder5>(
+            accountAndTransaction3_1_6.Links.links.GetStandingOrders,
+            accessToken
         )
 
         // Then
